Prevent corrupt messages from being added to the messageAdapter
In the chat, JsonDataExceptions causes old messages to be repeated.
This can happen when bots send controls that are structured incorrectly.
Since the error is thrown while updating the view holders, the view
recycling causes old values to persist.

By parsing the payload at an earlier stage we can prevent corrupt
messages from being added to the adapter list in the first place.